refactor: split patient registration into its own module; drop legacy registration.py
The doctor branch of the old registration.py is fully replaced by
api/signup.py + api/yoco.py. What's left is just patient self-registration,
which now lives in its own module so the file name reflects what it does.

The old on_user_verified hook is also gone — the only thing using it was
the doctor flow we just deleted.
